updated python script
git-svn-id: https://svn.code.sf.net/p/plexil/code/trunk@384 d2be67b0-814c-0410-949f-d215b6b29573

+++ bin/debugFlagCollector.py
@@ -203,8 +203,10 @@ def delete_temp_files():
 valid_plexil_home = check_env_variables()
 
 if valid_plexil_home:
-    os.remove(DEBUG_DEFINITIONS)
-    os.remove(COMPLETE_DEBUG)
+    if os.path.exists(DEBUG_DEFINITIONS):
+        os.remove(DEBUG_DEFINITIONS)
+    if os.path.exists(COMPLETE_DEBUG):
+        os.remove(COMPLETE_DEBUG)
 
     print "\nSearching for debug messages in .cc, .cpp, .hh and .h files...\n\n"
     locate_debug_msgs(FILE_TYPE="*.cc", PATTERN="ebugMsg(")
